# api/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.HTTPErrorDto:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  dtos.ServiceInstanceActionDto:
    properties:
      command:
        example: cmdExpose
        type: string
      form:
        example: ngx json form valid data
        type: string
      name:
        example: Expose service
        type: string
    type: object
  dtos.ServiceInstanceActionGroupDto:
    properties:
      actions:
        items:
          $ref: '#/definitions/dtos.ServiceInstanceActionDto'
        type: array
      name:
        example: Security
        type: string
    type: object
  dtos.ServiceInstanceDetailsDto:
    properties:
      actionGroups:
        items:
          $ref: '#/definitions/dtos.ServiceInstanceActionGroupDto'
        type: array
      name:
        example: my_kibana_instance_1
        type: string
      status:
        example: ok
        type: string
      type:
        example: kibana
        type: string
    type: object
  dtos.ServiceInstanceDetailsOverviewDto:
    properties:
      services:
        items:
          $ref: '#/definitions/dtos.ServiceInstanceDetailsDto'
        type: array
    type: object
...
  dtos.ServiceStoreItemYamlDto:
    properties:
      yaml:
        example: 'item: text'
        type: string
    type: object
  dtos.ServiceStoreOverviewDto:
    properties:
      services:
        items:
          $ref: '#/definitions/dtos.ServiceStoreItemDto'
        type: array
    type: object
  dtos.ServiceYamlDto:
    properties:
      yaml:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  description: Operator Automation Backend API overview.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Operator Automation Backend API
  version: "1.0"
paths:
...
  /services/{servicetype}/{servicename}:
    delete:
      consumes:
      - application/json
      description: Delete an instance of a service
      parameters:
      - description: Type of service
        in: path
        name: servicetype
        required: true
        type: string
      - description: Id of service
        in: path
        name: servicename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Delete a service instance
      tags:
      - Service
  /services/action/{servicetype}/{servicename}/{actioncommand}:
    post:
      consumes:
      - application/json
      description: Apply a service specific action to a service instance
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          type: string
      - description: Type of service
        in: path
        name: servicetype
        required: true
        type: string
      - description: Id of service
        in: path
        name: servicename
        required: true
        type: string
      - description: action command
        in: path
        name: actioncommand
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Apply a service specific action to a service instance
      tags:
      - Service
  /services/create/{servicetype}:
    post:
      consumes:
      - application/json
      description: Create an instance of a service from yaml
      parameters:
      - description: Service-Yaml
        in: body
        name: serviceyaml
        required: true
        schema:
          $ref: '#/definitions/dtos.ServiceYamlDto'
      - description: Type of service
        in: path
        name: servicetype
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Create service instance from yaml
      tags:
      - Service
  /services/info:
    get:
      consumes:
      - application/json
      description: Get an overview over all service instances
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ServiceInstanceDetailsOverviewDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Get an overview over all service instances
      tags:
      - Service
  /services/info/{servicetype}/{servicename}:
    get:
      consumes:
      - application/json
      description: Get details over a single service instance
      parameters:
      - description: Type of service
        in: path
        name: servicetype
        required: true
        type: string
      - description: Id of service
        in: path
        name: servicename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ServiceInstanceDetailsOverviewDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Get details over a single service instance
      tags:
      - Service
  /services/yaml/{servicetype}/{servicename}:
    get:
      consumes:
      - application/json
      description: Get the yaml file for an specific service instance. Parameter servicename
        has to be supplied.
      parameters:
      - description: Type of service
        in: path
        name: servicetype
        required: true
        type: string
      - description: Id of service
        in: path
        name: servicename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ServiceYamlDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Get the yaml file for an instance
      tags:
      - Service
  /servicestore/form/{servicetype}:
    get:
      consumes:
      - application/json
      description: Get the default yaml file for a service-template
      parameters:
      - description: Type of service
        in: path
        name: servicetype
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ServiceStoreItemYamlDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Get the json form for a service-template
      tags:
      - Servicestore
  /servicestore/info:
    get:
      consumes:
      - application/json
      description: Lists all possible deployable services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ServiceStoreOverviewDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Lists all possible deployable services
      tags:
      - Servicestore
  /servicestore/yaml/{servicetype}:
    post:
      consumes:
      - application/json
      description: Get the yaml for a service based on the filled form and the user
        data
      parameters:
      - description: Form-Result
        in: body
        name: formresult
        required: true
        schema:
          type: string
      - description: Type of service
        in: path
        name: servicetype
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ServiceStoreItemYamlDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.HTTPErrorDto'
      security:
      - BasicAuth: []
      summary: Get the yaml for a service
      tags:
      - Servicestore
schemes:
- https
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"
...